working openvz setup for debian squeeze

+++ setup.sh
@@ -23,12 +23,12 @@ base=/opt/farm/ext/openvz/templates/$OSVER
 if [ -d /srv/vz ]; then
 	echo "openvz already installed"
 	exit 0
-elif [ "$OSVER" != "debian-wheezy" ]; then
+elif [ "$OSVER" != "debian-squeeze" ] && [ "$OSVER" != "debian-wheezy" ]; then
 	echo "skipping openvz kernel setup, unsupported operating system version"
 	exit 1
 fi
 
-if [ ! -f /etc/apt/sources.list.d/proxmox.list ]; then
+if [ "$OSVER" = "debian-wheezy" ] && [ ! -f /etc/apt/sources.list.d/proxmox.list ]; then
 	echo "deb http://download.proxmox.com/debian wheezy pve-no-subscription" >/etc/apt/sources.list.d/proxmox.list
 	wget -O- "http://download.proxmox.com/debian/key.asc" |apt-key add -
 	apt-get update
@@ -46,12 +46,7 @@ if [ ! -d /srv/vz ]; then
 	mv /var/lib/vz /srv
 	ln -s /etc/vz /srv/vz/config
 
-	save_original_config /etc/init.d/vz
 	save_original_config /etc/vz/vz.conf
-	save_original_config /etc/vz/download.conf
-
-	sed -i -e "s/\(modify_vzconf\)$/\#\ \\1/" /etc/init.d/vz
-
 	set_openvz_option /etc/vz/vz.conf LOCKDIR '\/srv\/vz\/lock'
 	set_openvz_option /etc/vz/vz.conf DUMPDIR '\/srv\/vz\/dump'
 	set_openvz_option /etc/vz/vz.conf TEMPLATE '\/srv\/vz\/template'
@@ -59,7 +54,14 @@ if [ ! -d /srv/vz ]; then
 	set_openvz_option /etc/vz/vz.conf VE_PRIVATE '\/srv\/vz\/private\/\$VEID'
 	set_openvz_option /etc/vz/vz.conf IPTABLES '"ipt_REJECT ipt_tos ipt_limit ipt_multiport iptable_filter iptable_mangle ipt_TCPMSS ipt_tcpmss ipt_ttl ipt_length ipt_state"'
 	set_openvz_option /etc/vz/vz.conf IPV6 '"no"'
-	set_openvz_option /etc/vz/download.conf UPDATE_TEMPLATE '"no"'
 
-	install_copy $base/openvz-local.tpl /etc/sysctl.d/openvz-local.conf
+	if [ "$OSVER" = "debian-wheezy" ]; then
+		save_original_config /etc/vz/download.conf
+		set_openvz_option /etc/vz/download.conf UPDATE_TEMPLATE '"no"'
+
+		save_original_config /etc/init.d/vz
+		sed -i -e "s/\(modify_vzconf\)$/\#\ \\1/" /etc/init.d/vz
+
+		install_copy $base/openvz-local.tpl /etc/sysctl.d/openvz-local.conf
+	fi
 fi
